Add printing of 'const' qualifier.
Mark all tensors as const, then selectively
mark isConst as false for those that get written to.

This improves RAM usage on embedded targets significantly
(on GCC 8 at least): turns out GCC could not infere a initialized
tensor never written to to be constant. Now that they are const,
GCC can leave the contents in flash.

TODO: how to make an test case to check this?