Add Squeezenet 1.0 as performance test.
Contradictory to what the onnx model zoo page
suggest, SqueezeNet 1.1 is not 2.4x faster than
SqeezeNet 1.0 (at least the way we measure performance)
so this still is not a good perf measurement case.

Both run in ~100ms on a Intel(R) Core(TM) i5-7300U CPU,
when compiled with optimization and --fast-math.