Add AVR target and constant_access_code()
AVRs have limited RAM, but lots of flash (in comparison).
But they have the Harvard memory architecture, which means constants
are placed in RAM by default. This leads to RAM running out even with
small networks.

Because of all this, constant tensor accesses must be wrapped with a
macro that fetches the constant from instruction memory.

For now, ConvInteger and Gemm are "avr-ified".

For more info:
http://www.nongnu.org/avr-libc/user-manual/pgmspace.html

(I don't see an explanation on why the compiler can't insert that macro
by it self, once we have marked a tensor as PROGMEM. But it doesn't,
so onnx2c must do this kludge)