LSTM. Reset Y_c and Y_h at startup
Add a test that fails wihtout this.

Remove two simple tests that now fail. The exact reason is
not known, but the generation script for these is not the cleanest.
So assume that there was some error, since the new script
tests the exact same feature.

The old scripts were done with keras & tensorflow whereas the new
are sclbl & onnx-runtime. A worst-case scenario is that there is
a semantic error between the two frameworks. Unfortunatly
neiter gives detailed enough info on how exactly the LSTM should
be calculated...