Fix bug with out of order nodes.
The onnx graph does not need to list nodes in the order
they are resolvable. Up till now all graphs have had this order
though.

Add a out-of-order graph generated by sclblonnx as test case.

Implementation has a potential of being a bit slow (O(N^2)),
but that is ok, since it is a rare corner case, and N likely
is small for the onnx2c target audience.