Remove Node local copies of input and output
Having copies of the input and ouput tensors in a node
causes problems when trying to manipulate the graph.

Before now, there was no graph manipulating optimizations,
so this bad design could continue to be used. But now
it blocks the Cast-folding optimization pass.

Fix nodes up to (alphabetically) convinteger to comply.