Pull variable definitions out of for(...)
onnx2c defines counting variables for strings inside the for loops, like here:

for (unsigned i0=0; i0<1; i0++)

Unfortunately, older compilers do not accept this. In order to improve compatibility, variables need to be defined at the start of a function. This patch does this for the node types I encountered with my test data.